Fix lexer bug where the first body line wasn't properly MIME-decoded.
Reason for this was inappropriate readahead past the EOH \n character,
that caused input to be read in header mode when it was in fact body
and should have been subjected to MIME decoding.

We fix this by making sure that the EOH-producing rules
and related \n rules do not require readahead and by
forcing the scanner to always be interactive.

Tests have been adjusted after carefully verifying that
the lexer fix works as intended; tests had to be adjusted
because good.mbx misattributed the Reporting-MTA line to
a MIME header and prepended mime: when it was a body part.

git-svn-id: svn://svn.code.sf.net/p/bogofilter/code/trunk@6848 046ba1d7-8659-4f4f-8137-08d775d845e8