Cast toupper()/tolower() arguments to unsigned char.
This fixes compiler warnings, and prevents out-of-bounds array access when
characters between 0x80 and 0xff are encountered in configuration or command
line options or in MIME or "charset" headers.

A security audit of the affected 9 code places was made, and bogofilter
was found not to be vulnerable, because all occasions either process
trusted data, or data that was pre-validated by the lexer to either
match a lexer pattern or be purely "alnum" in the POSIX/C locale.

git-svn-id: svn://svn.code.sf.net/p/bogofilter/code/trunk@6876 046ba1d7-8659-4f4f-8137-08d775d845e8